Optionally use X-Forwarded-For in galette_logs (fixes #1031)
The IP address determination has been moved to its separate
function for easy further refactor.

The new integer configuration parameter works from the end of the
values of the X-Forwarded-For, i.e. it is the distance in the
chain of agents to the meaningful/trusted proxy.
Reminder:  prepending to the header is mostly trivial
(legitimately or not).